Use consul template in LWRP, bug fix in template
Template resources specified in the check and service providers didn't specify
the cookbook. So when the resource was used in a different cookbook, it tried
loading the template from the current cookbook (which doesn't exist unless the
cookbook was attempting to override the template).

This specifies the cookbook 'consul' in the LWRP template resources so that no
matter what cookbook uses the LWRPs, it will use the correct template.

Also fixes a small json bug in the service template.